chore(userspace/falco): simplify files naming
