+++ userspace/falco/outputs_http.cpp
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "falco_outputs_http.h"
+#include "outputs_http.h"
 #include "logger.h"
 #include "banned.h" // This raises a compilation error when certain functions are used
